Separate source files of Naive device functions.
