+++ primitiv/device_ops/copy_tensor_naive.cc
@@ -1,0 +1,20 @@
+#include <primitiv/config.h>
+
+#include <primitiv/naive_device.h>
+#include <primitiv/device_ops/naive_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Naive::copy_tensor_impl(const Tensor &x, Tensor &y) {
+  switch (x.device().type()) {
+    case Device::DeviceType::NAIVE:
+      reset_tensor_by_array(CDATA(x), y);
+      break;
+    default:
+      reset_tensor_by_vector(x.to_vector(), y);
+  }
+}
+
+}  // namespace devices
+}  // namespace primitiv
