+++ primitiv/device_ops/inplace_multiply_const_naive.cc
@@ -1,0 +1,16 @@
+#include <primitiv/config.h>
+
+#include <primitiv/naive_device.h>
+#include <primitiv/device_ops/naive_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Naive::inplace_multiply_const_impl(float k, Tensor &x) {
+  const std::uint32_t size = x.shape().size();
+  float *dest = MDATA(x);
+  REPEAT_OP(i, size, dest[i] *= k);
+}
+
+}  // namespace devices
+}  // namespace primitiv
