+++ primitiv/device_ops/random_bernoulli_naive.cc
@@ -1,0 +1,14 @@
+#include <primitiv/config.h>
+
+#include <primitiv/naive_device.h>
+#include <primitiv/device_ops/naive_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Naive::random_bernoulli_impl(float p, Tensor &y) {
+  randomizer_.fill_bernoulli(p, y.shape().size(), MDATA(y));
+}
+
+}  // namespace devices
+}  // namespace primitiv
