+++ primitiv/device_ops/random_log_normal_naive.cc
@@ -1,0 +1,14 @@
+#include <primitiv/config.h>
+
+#include <primitiv/naive_device.h>
+#include <primitiv/device_ops/naive_utils.h>
+
+namespace primitiv {
+namespace devices {
+
+void Naive::random_log_normal_impl(float mean, float sd, Tensor &y) {
+  randomizer_.fill_log_normal(mean, sd, y.shape().size(), MDATA(y));
+}
+
+}  // namespace devices
+}  // namespace primitiv
